Update notification config >= mitaka
Use oslo_messaging_notifications for mitaka or later releases
including setting the transport_url to the value provided by
the AMQP context.

This removes use of deprecated configuration options for
ceilometer notifications.

This change includes some refactoring to allow the topics to
use for notifications to be configured specifically for this
charm; future changes can use this to enable/disable designate
notifications dynamically.

Also includes redux of services check for amulet tests to drop
all checks apart from those for the neutron-api units.

Change-Id: Ib66371c0c479e0b341055941842e43ac57d4151d