Add neutron-load-balancer interface
The Neutron built-in LBaaS provider is deprecated as of
OpenStack version Queens and the service is to be replaced
by a separate service such as Octavia.

This interface serves the purpose of notifying a external
load balancer service of when the Neutron API is ready to
accept queries.

In a transition period it is also used by the ``neutron-api``
charm to determine whether it should configure Neutron with
the legacy LBaaS provider enabled or if it should enable
the ``lbaasv2-proxy`` driver to proxy load balancer requests
sent to the Neutron API to the external service.

Change-Id: Id9f7ffb3d363c7606d92af592b9803644046d865